Added slides & PDF
Presentation: No other sentence like ""Show me the money"" sparks more dread in MLOps teams & initiatives. And yet, the premise for most MLOps tools and teams is increased productivity which should lead to a healthier bottom-line.

In this talk I'll try to answer the following questions:

What are the North Star metrics of an effective ML and MLOps practice?
How do we connect the tactical, operational, & strategic layers of an MLOps practice?
How should MLOps and Data Science teams be organized to provide the most value to the organization?
We'll develop a potential framework using a combination of case studies of existing MLOps teams, evergreen wisdom from DevOps, & my own experiences traversing the ML lifecycle through a variety of industries & companies.